Move n+1 query test to api tests
